Predefined

Description

This module contains simple predefined colors and procedural texture functions.

Predefined colors

red :: Color

green :: Color

blue :: Color

white :: Color

black :: Color

grey :: Color

Procedural textures that take point return its material

Flat materials are 90% diffusive, shiny have 10% diffusion and 80% reflectance

flatRed :: Point3 -> Material

shinyRed :: Point3 -> Material

greenShinyRed :: Point3 -> Material

glass :: Point3 -> Material

Glass is fully reflective and refractive (conservation of energy is calculated from Fresnels law).

flatWhite :: Point3 -> Material

flatBlack :: Point3 -> Material

lightBulb :: Color -> Point3 -> Material

Creates a light material with specified color and 100% of emmision.

checkedMatt :: Point3 -> Material

Alternate 20x20x20 black and white cubes.

flatBlue :: Point3 -> Material

flatGreen :: Point3 -> Material

mirror :: Point3 -> Material