Raytracer

Description

This is an entry point to internals of Raytrace.

Documentation

raytrace

Arguments

:: Double
Importance of ray (when importance is sufficently low raytracing stops and returns black). Should be started at 1.0.

-> Int
Depth to trace, that is raytracing stops when depth reaches 0.

-> Ray
Ray to send.

-> ReaderT World (Rand StdGen) Color
Color of this ray.

Send a ray and return its color.

roots :: Double -> Double -> Double -> [Double]