Rendering

Description

This module contains functions used to render result image and save it.

Documentation

data Bitmap a

Generic Bitmap type with fast acces to it's size

pixels :: Bitmap a -> [a]

List of pixels of type a

width :: World -> Int

height :: World -> Int

savePPM

Arguments

:: FilePath
Path to a file

-> Bitmap Color
Bitmap to save

-> IO ()
Returned IO monad

Save color Bitmap to .pmm file.

renderAll_ :: Reader World (Bitmap Color)

Render all frames specified in World and return renderd bitmap of colors !! Will propabli become default rendering function in the future

renderAll

Arguments

:: String
Name of the file to save to

-> ReaderT World IO ()
Returned IO in ReaderT monad

Renders all frames specified in World and save them to a file

renderAllNotified

Arguments

:: String
Name of the file to save to

-> ReaderT World IO ()
Returned IO in ReaderT monad

Render all frames specified in World and save them to file, notifing each time a frame is rendered.

Orphan instances

NFData Vec3

Methods

rnf :: Vec3 -> ()

NFData Normal3

Methods

rnf :: Normal3 -> ()

NFData Ray

Methods

rnf :: Ray -> ()