SceneBuilder

Description

This module contains functions used to build and manipulate everything that is to be rendered. They should be used in state monad as with modify function, however in case of different need they are defined as pure modifiers.

Documentation

emptyScene :: Scene

Just an empty scene without any objects, with grey ambient light and black background.

addShape

Arguments

:: Shape
Shape that has to be added

-> Scene

-> Scene

Adds shape to a scene

addLight

Arguments

:: Point3
Center of a light source

-> Double
Radius of a light sphere

-> Color
Color of a light

-> Scene

-> Scene

Adds spherical light to a scene

setAmbientLight

Arguments

:: Color
New ambient light

-> Scene

-> Scene

Changes ambient light in a scene

setBackgroundColor

Arguments

:: Color
New background color

-> Scene

-> Scene

Changes background color of a scene

sampleView :: View

Sample view with camera at (0,0,0), looking in (0,0,1) direction, oriented along y and with field of view equal 110

yawCamera

Arguments

:: Double
Number of radians to rotate by

-> View

-> View

Yaws camera forward, that is rotates around vertical/right vector

positionCamera

Arguments

:: Vec3
New camera position

-> View

-> View

Moves camera to a new position

sampleWorld :: World

Sample World with emptyScene and sampleView with ray depth of 10, 1 frame to render and resolution of 500x500

setScene

Arguments

:: Scene
Scene to save

-> World

-> World

Save Scene to a World

setView

Arguments

:: View
New View to save

-> World

-> World

Save View to a World

setRayDepth

Arguments

:: Int
New depth of rays

-> World

-> World

Set depth of rays in a World

setSamples

Arguments

:: Int
Number of samples

-> World

-> World

Set number of samples to average when rendering this World

setWidth :: Int -> World -> World

Set width of rendered image

setHeight :: Int -> World -> World

Set height of rendered image