Types

Description

Module containing types used in Raytace library.

Documentation

type Color = Vec3

clamp :: Color -> Color

addMix :: Color -> Color -> Color

subMix :: Color -> Color -> Color

prodMix :: Color -> Color -> Color

type Point3 = Vec3

data BRDF

BRDF models with their corresponding descriptions (Bounce Reflect Diffuse Functions).

Constructors

Lambert
Lambertian reflectance BRDF.

Fields

color :: Color

Color of lambertian surface.

Reflection
Perfect reflection BRDF.

Fields

inRI :: Double

Index of refraction of material inside the shape.

outRI :: Double

Index of refraction of material outside the shape.

color :: Color

Color of lambertian surface.

Refraction
Perfect refraction BRDF.

Fields

inRI :: Double

Index of refraction of material inside the shape.

outRI :: Double

Index of refraction of material outside the shape.

color :: Color

Color of lambertian surface.

Emmisive
Generic emmisive BRDF.

Fields

color :: Color

Color of lambertian surface.

data Material

Material of specific point.

Constructors

Material

Fields

brdfs :: [(Double, BRDF)]

Pairs of and brdfs. In order for material to be realistic sum of must be lower or equal to 1.

data Ray

Ray in three dimensions.

Constructors

Ray

Fields

base :: Point3

Point from which the ray starts.

direction :: Normal3

Direction in which ray is pointing.

data Intersection

All information about intersection.

Constructors

Intersection

Fields

normalI :: Normal3

Normal of intersection.

point :: Point3

Point of intersection.

ray :: Ray

Ray which intersected.

material :: Material

Material at point of intersection.

data Scene

Scenes description.

Constructors

Scene

Fields

shapes :: [Shape]

Shapes in a scene.

lights :: [Light]

Lights in a scene.

ambientLight :: Color

Ambient light of a scene (light that doesn't depend on visibility from any light source).

backgroundColor :: Color

data View

View description.

Constructors

View

Fields

camera :: Point3

Cameras position in the scene

clipPlane :: Double

Clipping plane of a camera (how close the objects can be seen)

forward :: Normal3

Normal vector of where the camera is facing

up :: Normal3

Normal vector of where the camera has "up" direction in

fov :: Double

Angle between point most to the left and most to the right, as seen by the camera

data World

Description of whole Raytrace world with all settings needed to render the scene.

Constructors

World !Scene !View !Int !Int !Int !Int
Scene to be rendered, view from which to render, depth of ray bounces, number of frames to be averaged width and height of output render

scene :: World -> Scene

view :: World -> View

rayDepth :: World -> Int

samples :: World -> Int

width :: World -> Int

height :: World -> Int

data Shape

Shapes in the scene

Constructors

Sphere
Spherical shape.

Fields

center :: Point3

Center of a sphere.

radius :: Double

Radius of a sphere.

materialFunc :: Point3 -> Material

Function that describes material of point on a sphere.

Plane
Shape as two sided infinte plane.

Fields

normal :: Normal3

Normal direction of a sphere.

distance :: Double

Distance from an origin - (0,0,0) along normal.

materialFunc :: Point3 -> Material

Function that describes material of point on a sphere.

type Light = Point3

Light point used as general check if point is lit

Orphan instances

Eq Vec3

Methods

(==) :: Vec3 -> Vec3 -> Bool

(/=) :: Vec3 -> Vec3 -> Bool

Eq Normal3

Methods

(==) :: Normal3 -> Normal3 -> Bool

(/=) :: Normal3 -> Normal3 -> Bool